feat(document): change admins and some models for admins api

+++ src/routes/admins.ts
@@ -1,44 +1,70 @@
 import express, { Request } from "express";
 import db from "../connection/connection";
+import { auth } from "../middlewares/auth/auth";
 import { Favorite } from "../models/favorites.model";
 import { Movie } from "../models/movies.model";
+import { User } from "../models/user.model";
 import { formatCollection } from "../utils";
 import router from "./movies";
 
-router.post(`/films`, async ({ body: movie }: Request<{}, {}, Movie>, res) => {
+router.post(`/:username/movies`, auth, async ({ body: movie, params: {username} }: Request<Partial<User>, {}, Movie, {}>, res) => {
     const movies = formatCollection<Movie>(await db.collection("Films").get());
-    const hasSameType = JSON.stringify(Object.keys(movie).sort()) === JSON.stringify(Object.keys(movies[0]).sort());
-    const isIdPresent = movies.some(({ id }) => id === movie.id);
-    if (hasSameType) {
-        if (isIdPresent) {
-            res.status(400).json({ error: 400, message: `Operation blocked, ID: ${movie.id} is already present` })
-        } else {
-            const docRef = db.collection('Films').doc(String(movie.id));
+    const users = formatCollection<User>(await db.collection("Users").get());
+    const index = users.findIndex(({ username: uUsername, admin }) => (admin==true && uUsername === username));
+    if(index>-1) {
+        let max = Math.max(...movies.map(({ id }) => Number(id)) as number[]) + 1;
+        let newMovie = {...movie, id: max}
+        const hasSameType = JSON.stringify(Object.keys(newMovie).sort()) === JSON.stringify(Object.keys(movies[0]).sort());
+        if (hasSameType) {
+            const docRef = db.collection('Films').doc(String(max));
             await docRef.set({
                 movie
             })
             res.status(200).json({ message: `Movie inserted correctly!`, movie: movie });
+        } else {
+            res.status(400).json({ error: 400, message: `Operation blocked, Please fill in all the required fields` })
         }
-    } else {
-        res.status(400).json({ error: 400, message: `Operation blocked, Please fill in all the required fields` })
+    }else {
+        res.status(403).json({error: "403", message: "user nor authorized"})
     }
 
 })
 
-router.delete(`/films/:movieId`, async ({ params: { movieId } }: Request<{ movieId: string }, {}, {}, {}>, res) => {
+router.delete(`/:username/movies/:movieId`, auth ,async ({ params: { movieId, username } }: Request<{ movieId: string, username: string }, {}, {}, {}>, res) => {
     const movies = formatCollection<Movie>(await db.collection("Films").get());
-    const index = movies.findIndex(({ id }) => id === movieId);
-    if (index > -1) {
-        const favorites = formatCollection<Favorite>(await db.collection("Favorites").get());
-        const indexFavorite = favorites.findIndex(({ movieId: favMovieId }) => String(favMovieId) === movieId);
-        if(indexFavorite > -1){
-            db.collection('Favorites').doc(favorites[indexFavorite].id).delete();
+    const users = formatCollection<User>(await db.collection("Users").get());
+    const userIndex = users.findIndex(({ username: uUsername, admin }) => (admin==true && uUsername === username));
+    if(userIndex>-1) {
+        const index = movies.findIndex(({ id }) => id === movieId);
+        if (index > -1) {
+            const favorites = formatCollection<Favorite>(await db.collection("Favorites").get());
+            const indexFavorite = favorites.findIndex(({ movieId: favMovieId }) => String(favMovieId) === movieId);
+            if(indexFavorite > -1){
+                db.collection('Favorites').doc(favorites[indexFavorite].id).delete();
+            }
+            db.collection('Films').doc(movies[index].id).delete();
+            return res.status(200).json({ message: `Movie with ID: ${movies[index].id} removed` })
         }
-        db.collection('Films').doc(movies[index].id).delete();
-        return res.status(200).json({ message: `Movie with ID: ${movies[index].id} removed` })
+            return res.json({ error: "404", message: `Movie with ID: ${movieId} not found` })
+    }else {
+        res.status(403).json({error: "403", message: "user nor authorized"});
     }
-        return res.json({ error: "404", message: `Movie with ID: ${movieId} not found` })
-    
+})
+
+router.put('/:username/user/:username1', auth, async({params: {username, username1} }: Request<{username: string, username1: string} , {}, {}, {}>, res)=>{
+    const users = formatCollection<User>(await db.collection("Users").get());
+    const userIndex = users.findIndex(({ username: uUsername, admin }) => (admin==true && uUsername === username));
+    if(userIndex>-1){
+        const user = users.find(({username})=>username===username1);
+        if(user) {
+            db.collection("Users").doc(user.id).update({admin: true});
+            res.json({message: "new admin setted"});
+        }else {
+            res.status(404).json({error: "404", message: "user not found"});
+        }
+    } else {
+        res.status(403).json({error: "403", message: "user nor authorized"});
+    }
 })
 
 export default router;
